requirements.txt
gradio==3.50.2
python-telegram-bot==20.6
httpx==0.25.2
fastapi==0.100.0
pydantic==1.10.12
scikit-learn==1.4.2
joblib==1.3.2
nest_asyncio==1.6.0
uvicorn==0.23.2
